Add task management code with first pattern
Add task management code to manage timing of arbitrary tasks to be run
in a round-robin outer-most control loop (polymorphic). Integrate
the first tested (demo) pattern into this as a pattern to be run
within a "pattern runner" task. Other tasks will be created for
checking the battery and polling the user button, for example.