Final tweaks before the wedding!
Added an all dark pattern so that we can always start from a known
state. On the dress, unlike on the tester platform, some pixels start
stuck on. It was helpful to start it, get past this uninitialized
state by getting back to the dark pattern, then doing the ramp to
white when we kiss at the ceremony.